+++ src/lib.rs
@@ -683,10 +683,7 @@
 //!     table.to_string(PostgresQueryBuilder),
 //!     r#"TRUNCATE TABLE "font""#
 //! );
-//! assert_eq!(
-//!     table.to_string(SqliteQueryBuilder),
-//!     r#"TRUNCATE TABLE "font""#
-//! );
+//! // Sqlite does not support the TRUNCATE statement
 //! ```
 //!
 //! ### Foreign Key Create