+++ tests/sqlite/table.rs
@@ -329,6 +329,7 @@ fn drop_1() {
 }
 
 #[test]
+#[should_panic(expected = "Sqlite doesn't support TRUNCATE statement")]
 fn truncate_1() {
     assert_eq!(
         Table::truncate()